requirements.txt
streamlit
google-genai
requests
gTTS
duckduckgo-search